More agent test moves
