+++ packages/core/src/agent/agent.test.ts
@@ -18,8 +18,6 @@ import { Mastra } from '../mastra';
 import type { MastraMessageV2, StorageThreadType } from '../memory';
 import { MockMemory } from '../memory/mock';
 import { RuntimeContext } from '../runtime-context';
-import { createScorer } from '../scores';
-import { runScorer } from '../scores/hooks';
 import { MockStore } from '../storage';
 import type { MastraModelOutput } from '../stream/base/output';
 import { createTool } from '../tools';
@@ -44,10 +42,6 @@ const mockFindUser = vi.fn().mockImplementation(async data => {
   if (!userInfo) return { message: 'User not found' };
   return userInfo;
 });
-
-vi.mock('../scores/hooks', () => ({
-  runScorer: vi.fn(),
-}));
 
 const openai = createOpenAI({ apiKey: process.env.OPENAI_API_KEY });
 const openai_v5 = createOpenAIV5({ apiKey: process.env.OPENAI_API_KEY });
@@ -5861,294 +5855,6 @@ function agentTests({ version }: { version: 'v1' | 'v2' }) {
     });
   });
 
-  describe(`${version} - UIMessageWithMetadata support`, () => {
-    let dummyModel: MockLanguageModelV1 | MockLanguageModelV2;
-    const mockMemory = new MockMemory();
-
-    beforeEach(() => {
-      if (version === 'v1') {
-        dummyModel = new MockLanguageModelV1({
-          doGenerate: async () => ({
-            finishReason: 'stop',
-            usage: { completionTokens: 10, promptTokens: 3 },
-            text: 'Response acknowledging metadata',
-            rawCall: { rawPrompt: null, rawSettings: {} },
-          }),
-          doStream: async () => ({
-            stream: simulateReadableStream({
-              chunks: [
-                { type: 'text-delta', textDelta: 'Response' },
-                { type: 'text-delta', textDelta: ' acknowledging' },
-                { type: 'text-delta', textDelta: ' metadata' },
-                {
-                  type: 'finish',
-                  finishReason: 'stop',
-                  logprobs: undefined,
-                  usage: { completionTokens: 10, promptTokens: 3 },
-                },
-              ],
-            }),
-            rawCall: { rawPrompt: null, rawSettings: {} },
-          }),
-        });
-      } else {
-        dummyModel = new MockLanguageModelV2({
-          doStream: async () => ({
-            stream: convertArrayToReadableStream([
-              { type: 'stream-start', warnings: [] },
-              { type: 'response-metadata', id: 'id-0', modelId: 'mock-model-id', timestamp: new Date(0) },
-              { type: 'text-start', id: '1' },
-              { type: 'text-delta', id: '1', delta: 'Response' },
-              { type: 'text-delta', id: '1', delta: ' acknowledging' },
-              { type: 'text-delta', id: '1', delta: ' metadata' },
-              { type: 'text-end', id: '1' },
-              { type: 'finish', finishReason: 'stop', usage: { inputTokens: 10, outputTokens: 10, totalTokens: 20 } },
-            ]),
-            rawCall: { rawPrompt: null, rawSettings: {} },
-            warnings: [],
-          }),
-        });
-      }
-    });
-
-    it('should preserve metadata in generate method', async () => {
-      const agent = new Agent({
-        name: 'metadata-test-agent',
-        instructions: 'You are a helpful assistant',
-        model: dummyModel,
-        memory: mockMemory,
-      });
-
-      const messagesWithMetadata = [
-        {
-          role: 'user' as const,
-          content: 'Hello with metadata',
-          parts: [{ type: 'text' as const, text: 'Hello with metadata' }],
-          metadata: {
-            source: 'web-ui',
-            customerId: '12345',
-            context: { orderId: 'ORDER-789', status: 'pending' },
-          },
-        },
-      ];
-
-      if (version === 'v1') {
-        await agent.generateLegacy(messagesWithMetadata, {
-          memory: {
-            resource: 'customer-12345',
-            thread: {
-              id: 'support-thread',
-            },
-          },
-        });
-      } else {
-        await agent.generate(messagesWithMetadata, {
-          memory: {
-            resource: 'customer-12345',
-            thread: {
-              id: 'support-thread',
-            },
-          },
-        });
-      }
-      // Verify messages were saved with metadata
-      const savedMessages = await mockMemory.getMessages({
-        threadId: 'support-thread',
-        resourceId: 'customer-12345',
-        format: 'v2',
-        selectBy: {
-          last: 10,
-        },
-      });
-
-      expect(savedMessages.length).toBeGreaterThan(0);
-
-      // Find the user message
-      const userMessage = savedMessages.find(m => m.role === 'user');
-      expect(userMessage).toBeDefined();
-
-      // Check that metadata was preserved in v2 format
-      if (
-        userMessage &&
-        'content' in userMessage &&
-        typeof userMessage.content === 'object' &&
-        'metadata' in userMessage.content
-      ) {
-        expect(userMessage.content.metadata).toEqual({
-          source: 'web-ui',
-          customerId: '12345',
-          context: { orderId: 'ORDER-789', status: 'pending' },
-        });
-      }
-    });
-
-    it('should preserve metadata in stream method', async () => {
-      const agent = new Agent({
-        name: 'metadata-stream-agent',
-        instructions: 'You are a helpful assistant',
-        model: dummyModel,
-        memory: mockMemory,
-      });
-
-      const messagesWithMetadata = [
-        {
-          role: 'user' as const,
-          content: 'Stream with metadata',
-          parts: [{ type: 'text' as const, text: 'Stream with metadata' }],
-          metadata: {
-            source: 'mobile-app',
-            sessionId: 'session-123',
-            deviceInfo: { platform: 'iOS', version: '17.0' },
-          },
-        },
-      ];
-
-      let stream;
-      if (version === 'v1') {
-        stream = await agent.streamLegacy(messagesWithMetadata, {
-          memory: {
-            resource: 'user-mobile',
-            thread: {
-              id: 'mobile-thread',
-            },
-          },
-        });
-      } else {
-        stream = await agent.stream(messagesWithMetadata, {
-          memory: {
-            resource: 'user-mobile',
-            thread: {
-              id: 'mobile-thread',
-            },
-          },
-        });
-      }
-
-      // Consume the stream
-      let finalText = '';
-      for await (const textPart of stream.textStream) {
-        finalText += textPart;
-      }
-
-      expect(finalText).toBe('Response acknowledging metadata');
-
-      // Verify messages were saved with metadata
-      const savedMessages = await mockMemory.getMessages({
-        threadId: 'mobile-thread',
-        resourceId: 'user-mobile',
-        format: 'v2',
-        selectBy: {
-          last: 10,
-        },
-      });
-
-      expect(savedMessages.length).toBeGreaterThan(0);
-
-      // Find the user message
-      const userMessage = savedMessages.find(m => m.role === 'user');
-      expect(userMessage).toBeDefined();
-
-      // Check that metadata was preserved
-      if (
-        userMessage &&
-        'content' in userMessage &&
-        typeof userMessage.content === 'object' &&
-        'metadata' in userMessage.content
-      ) {
-        expect(userMessage.content.metadata).toEqual({
-          source: 'mobile-app',
-          sessionId: 'session-123',
-          deviceInfo: { platform: 'iOS', version: '17.0' },
-        });
-      }
-    });
-
-    it('should handle mixed messages with and without metadata', async () => {
-      const agent = new Agent({
-        name: 'mixed-metadata-agent',
-        instructions: 'You are a helpful assistant',
-        model: dummyModel,
-        memory: mockMemory,
-      });
-
-      const mixedMessages = [
-        {
-          role: 'user' as const,
-          content: 'First message with metadata',
-          parts: [{ type: 'text' as const, text: 'First message with metadata' }],
-          metadata: {
-            messageType: 'initial',
-            priority: 'high',
-          },
-        },
-        {
-          role: 'assistant' as const,
-          content: 'Response without metadata',
-          parts: [{ type: 'text' as const, text: 'Response without metadata' }],
-        },
-        {
-          role: 'user' as const,
-          content: 'Second user message',
-          parts: [{ type: 'text' as const, text: 'Second user message' }],
-          // No metadata on this message
-        },
-      ];
-
-      if (version === 'v1') {
-        await agent.generateLegacy(mixedMessages, {
-          memory: {
-            resource: 'mixed-user',
-            thread: {
-              id: 'mixed-thread',
-            },
-          },
-        });
-      } else {
-        await agent.generate(mixedMessages, {
-          memory: {
-            resource: 'mixed-user',
-            thread: {
-              id: 'mixed-thread',
-            },
-          },
-        });
-      }
-      // Verify messages were saved correctly
-      const savedMessages = await mockMemory.getMessages({
-        threadId: 'mixed-thread',
-        resourceId: 'mixed-user',
-        format: 'v2',
-        selectBy: {
-          last: 10,
-        },
-      });
-
-      expect(savedMessages.length).toBeGreaterThan(0);
-
-      // Find messages and check metadata
-      const messagesAsV2 = savedMessages as MastraMessageV2[];
-      const firstUserMessage = messagesAsV2.find(
-        m =>
-          m.role === 'user' &&
-          m.content.parts?.[0]?.type === 'text' &&
-          m.content.parts[0].text.includes('First message'),
-      );
-      const secondUserMessage = messagesAsV2.find(
-        m =>
-          m.role === 'user' && m.content.parts?.[0]?.type === 'text' && m.content.parts[0].text.includes('Second user'),
-      );
-
-      // First message should have metadata
-      expect(firstUserMessage?.content.metadata).toEqual({
-        messageType: 'initial',
-        priority: 'high',
-      });
-
-      // Second message should not have metadata
-      expect(secondUserMessage?.content.metadata).toBeUndefined();
-    });
-  });
-
   it(`${version} - stream - should pass and call client side tools with experimental output`, async () => {
     const userAgent = new Agent({
       name: 'User agent',
@@ -6301,252 +6007,6 @@ function agentTests({ version }: { version: 'v1' | 'v2' }) {
 
       expect(agentToolCall).toBeDefined();
     }, 50000);
-  });
-
-  describe('scorer output data', () => {
-    it(`${version} - should return scoring data from generate when returnScorerData is true`, async () => {
-      const agent = new Agent({
-        name: 'Scorer Agent',
-        instructions: 'You are an agent that can score things',
-        model: dummyModel,
-      });
-
-      let result;
-      if (version === 'v1') {
-        result = await agent.generateLegacy('Make it green', {
-          returnScorerData: true,
-        });
-      } else {
-        result = await agent.generate('Make it green', {
-          returnScorerData: true,
-        });
-      }
-
-      expect(result.scoringData).toBeDefined();
-      expect(result.scoringData.input).toMatchObject({
-        inputMessages: expect.any(Array),
-        rememberedMessages: expect.any(Array),
-        systemMessages: expect.any(Array),
-        taggedSystemMessages: expect.any(Object),
-      });
-      expect(result.scoringData.output).toBeInstanceOf(Array);
-    });
-
-    it(`${version} - should not return scoring data from generate when returnScorerData is false`, async () => {
-      const agent = new Agent({
-        name: 'Scorer Agent',
-        instructions: 'You are an agent that can score things',
-        model: dummyModel,
-      });
-
-      let result;
-      if (version === 'v1') {
-        result = await agent.generateLegacy('Make it green', {
-          returnScorerData: false,
-        });
-      } else {
-        result = await agent.generate('Make it green', {
-          returnScorerData: false,
-        });
-      }
-
-      expect(result.scoringData).toBeUndefined();
-    });
-
-    it(`${version} - should not return scoring data from generate when returnScorerData is not specified`, async () => {
-      const agent = new Agent({
-        name: 'Scorer Agent',
-        instructions: 'You are an agent that can score things',
-        model: dummyModel,
-      });
-
-      let result;
-      if (version === 'v1') {
-        result = await agent.generateLegacy('Make it green');
-      } else {
-        result = await agent.generate('Make it green');
-      }
-
-      expect(result.scoringData).toBeUndefined();
-    });
-  });
-
-  describe('scorer override functionality', () => {
-    let agent: Agent;
-    let mastra: Mastra;
-    let scorerTest: any;
-    let scorer1: any;
-
-    beforeEach(() => {
-      vi.clearAllMocks();
-      scorerTest = createScorer({
-        name: 'scorerTest',
-        description: 'Test Scorer',
-      }).generateScore(() => 0.95);
-
-      scorer1 = createScorer({
-        name: 'scorer1',
-        description: 'Test Scorer 1',
-      }).generateScore(() => 0.95);
-
-      agent = new Agent({
-        name: 'Test Agent',
-        instructions: 'You are a test agent.',
-        model: dummyModel,
-        scorers: {
-          scorerTest: {
-            scorer: scorerTest,
-          },
-        },
-      });
-
-      mastra = new Mastra({
-        agents: { agent },
-        logger: false,
-        scorers: { scorer1 },
-      });
-    });
-
-    it(`${version} - should call scorerTest when no override is provided`, async () => {
-      if (version === 'v1') {
-        await agent.generateLegacy('Hello world');
-      } else {
-        await agent.generate('Hello world');
-      }
-
-      expect(runScorer).toHaveBeenCalledWith(
-        expect.objectContaining({
-          scorerId: 'scorerTest',
-          scorerObject: expect.objectContaining({
-            scorer: scorerTest,
-          }),
-        }),
-      );
-    });
-
-    it(`${version} - should use override scorers when provided in generate options`, async () => {
-      if (version === 'v1') {
-        await agent.generateLegacy('Hello world', {
-          scorers: {
-            scorer1: { scorer: mastra.getScorer('scorer1') },
-          },
-        });
-      } else {
-        await agent.generate('Hello world', {
-          scorers: {
-            scorer1: { scorer: mastra.getScorer('scorer1') },
-          },
-        });
-      }
-
-      expect(runScorer).toHaveBeenCalledWith(
-        expect.objectContaining({
-          scorerId: 'scorer1',
-          scorerObject: expect.objectContaining({
-            scorer: expect.any(Object),
-          }),
-        }),
-      );
-
-      expect(runScorer).not.toHaveBeenCalledWith(
-        expect.objectContaining({
-          scorerId: 'scorerTest',
-          scorerObject: expect.objectContaining({
-            scorer: scorerTest,
-          }),
-        }),
-      );
-
-      expect(runScorer).toHaveBeenCalledTimes(1);
-    });
-
-    it(`${version} - should call scorers when provided in stream options`, async () => {
-      let result: any;
-      if (version === 'v1') {
-        result = await agent.streamLegacy('Hello world', {
-          scorers: {
-            scorer1: { scorer: mastra.getScorer('scorer1') },
-          },
-        });
-      } else {
-        result = await agent.stream('Hello world', {
-          scorers: {
-            scorer1: { scorer: mastra.getScorer('scorer1') },
-          },
-        });
-      }
-      await result.consumeStream();
-
-      expect(runScorer).toHaveBeenCalledWith(
-        expect.objectContaining({
-          scorerId: 'scorer1',
-          scorerObject: expect.objectContaining({
-            scorer: expect.any(Object),
-          }),
-        }),
-      );
-    });
-
-    it(`${version} - can use scorer name for scorer config for generate`, async () => {
-      if (version === 'v1') {
-        await agent.generateLegacy('Hello world', {
-          scorers: {
-            scorer1: { scorer: scorer1.name },
-          },
-        });
-      } else {
-        await agent.generate('Hello world', {
-          scorers: {
-            scorer1: { scorer: scorer1.name },
-          },
-        });
-      }
-
-      expect(runScorer).toHaveBeenCalledWith(
-        expect.objectContaining({
-          scorerId: 'scorer1',
-          scorerObject: expect.objectContaining({
-            scorer: scorer1,
-          }),
-        }),
-      );
-    });
-
-    it(`${version} - should call runScorer with correct parameters`, async () => {
-      if (version === 'v1') {
-        await agent.generateLegacy('Hello world', {
-          scorers: {
-            scorer1: { scorer: scorer1.name },
-          },
-        });
-      } else {
-        await agent.generate('Hello world', {
-          scorers: {
-            scorer1: { scorer: scorer1.name },
-          },
-        });
-      }
-
-      // Verify the exact call parameters
-      expect(runScorer).toHaveBeenCalledWith({
-        scorerId: 'scorer1',
-        scorerObject: { scorer: scorer1 },
-        runId: expect.any(String),
-        input: expect.any(Object),
-        output: expect.any(Object),
-        runtimeContext: expect.any(Object),
-        entity: expect.objectContaining({
-          id: 'Test Agent',
-          name: 'Test Agent',
-        }),
-        source: 'LIVE',
-        entityType: 'AGENT',
-        structuredOutput: false,
-        threadId: undefined,
-        resourceId: undefined,
-        tracingContext: expect.any(Object),
-      });
-    });
   });
 
   describe('defaultStreamOptions onFinish callback bug', () => {
