+++ packages/server/src/server/handlers/memory.test.ts
@@ -1,10 +1,9 @@
 import { Agent } from '@mastra/core/agent';
 import type { CoreMessage } from '@mastra/core/llm';
 import { Mastra } from '@mastra/core/mastra';
-import type { MastraMessageV1, MastraMessageV2 } from '@mastra/core/memory';
 import { MockMemory } from '@mastra/core/memory';
+import type { MastraMessageV1, MastraMessageV2, StorageThreadType } from '@mastra/core/memory';
 import { InMemoryStore } from '@mastra/core/storage';
-import type { StorageThreadType } from '@mastra/core/memory';
 import type { Mock } from 'vitest';
 import { describe, it, expect, vi, beforeEach } from 'vitest';
 import { HTTPException } from '../http-exception';
